+++ src/app/components/users/users.component.ts
@@ -1,5 +1,6 @@
-import { Component, OnInit,NgZone } from '@angular/core';
+import { Component, OnInit, NgZone } from '@angular/core';
 import { Router } from '@angular/router';
+import { CrudService } from 'src/app/service/crud.service';
 
 @Component({
   selector: 'app-users',
@@ -8,21 +9,56 @@ import { Router } from '@angular/router';
 })
 export class UsersComponent implements OnInit {
 
-  isExpand:boolean=false;
+  isExpand: boolean = true;
   constructor(
-    private ngZone:NgZone,
-    private router:Router
+    private ngZone: NgZone,
+    private router: Router,
+    private crudService: CrudService
   ) { }
 
   ngOnInit(): void {
+    this.getUsername();
+    this.getUserRole();
   }
 
-  logout(){
-    alert("are you sure want to logout.....?");
+  logout() {
+    if (window.confirm('Do you want to Logout?')) {
     sessionStorage.clear();
     this.ngZone.run(() => this.router.navigateByUrl('/login'))
     console.log("user logout successfull...");
-    
+    }    
   }
+
+  username!: string | null;
+  getUsername(): void {
+
+    this.username = sessionStorage.getItem('username');
+    console.log('username ' + this.username);
+
+
+  }
+
+  //Get Users Role
+  User: any = [];
+  role!: string;
+  getUserRole() {
+    this.crudService.GetUsers().subscribe((res) => {
+      this.User = res;
+      console.log(this.User)
+
+      //iterate through array
+      this.User.forEach((value: any) => {
+
+        if (value.Username == this.username) {
+          this.role = value.Role;
+          console.log(this.role);
+        }
+
+      });
+
+    });
+
+  }
+
 
 }
